Split manifest hook handling
